Update to refactor naming of taxonomy-related module.
